Delete remote chunks folder when canceling upload
